[PTZ] Put zoom behind the permission on Desktop
This CL makes sure zoom is put behind the panTiltZoom permission like
pan and tilt. The only exception is on Android where zoom is granted
automatically.

This CL also adds proper PTZ camera support on Android so that web
developers who specifically ask for zoom on an Android camera that does
not support it (e.g. infrared) get an OverConstrainedError.

Change-Id: I80a37e75d6eec7ab134b9296c222a6d450f26078
Bug: 1104080